requirements.txt
cryptography==42.0.5
fastapi==0.110.0


langchain_community==0.0.28
langchain_core==0.1.31
langchain_openai==0.0.8
openai==1.14.0

python-dotenv==1.0.1
SQLAlchemy==2.0.25
chromadb==0.4.22
beautifulsoup4==4.12.3
PyMySQL==1.1.0


aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2

bleach==6.1.0
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3

click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1

dataclasses-json==0.6.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1

fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
groq==0.4.2
grpcio==1.60.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kubernetes==29.0.0
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-groq==0.0.1
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.40
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0

opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
posthog==3.4.1
prompt-toolkit==3.0.43
protobuf==4.25.2
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2

PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2

pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5

stack-data==0.6.3
starlette==0.36.3
streamlit==1.31.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0